+++ src/cli/utils.rs
@@ -4,6 +4,7 @@ use chrono::NaiveDateTime;
 use itertools::Itertools;
 use serde::Serialize;
 use std::collections::{HashMap, HashSet};
+use std::path::Path;
 
 #[derive(Serialize, tabled::Tabled)]
 pub struct CollectorInfo {
@@ -54,4 +55,28 @@ pub fn get_missing_collectors(latest_items: &[BrokerItem]) -> Vec<CollectorInfo>
         .collect();
 
     missing_collectors
+}
+
+pub fn is_local_path(path: &str) -> bool {
+    if path.contains("://") {
+        return false;
+    }
+    let path = Path::new(path);
+    path.is_absolute() || path.is_relative()
+}
+
+pub fn parse_s3_path(path: &str) -> Option<(String, String)> {
+    // split a path like s3://bucket/path/to/file into (bucket, path/to/file)
+    let parts = path.split("://").collect::<Vec<&str>>();
+    if parts.len() != 2 || parts[0] != "s3" {
+        return None;
+    }
+    let parts = parts[1].split('/').collect::<Vec<&str>>();
+    let bucket = parts[0].to_string();
+    // join the rest delimited by `/`
+    let path = format!("/{}", parts[1..].join("/"));
+    if parts.ends_with(&["/"]) {
+        return None;
+    }
+    Some((bucket, path))
 }